Custom validations
Added validation for fields that rely on javascript and cannot be automatically validated by bootstrap (ingredients and method lists). Slightly tweaked schemas in a way that allows for bootstrap validation to work for preparation time and serves. As bootstrap validation affects spacing I also adjusted the width of some input fields to ensure a smooth user experience. Lastly, some guidance wording was added on the new and edit page to facilitate users with inputing data.