Styling Application
Several enhancements were made to the application's interface and functionality. These include a redesign of the home page layout, ensuring uniformity in the height of index cards, and implementing a default display of metric measurements on show pages. The layout of show pages was also improved for better readability and organization, irrespective of user login status. Additionally, the login and signup pages' styles were updated, and a general.css document was set up to maintain consistent styles across the app. Lastly, redundant navigation links were removed across the application to streamline navigation and enhance the overall user experience.